More work on dialog integration

+++ GameCore/GUI/GameDialog/LineData.cs
@@ -2,8 +2,8 @@
 
 public class LineData : IStatement
 {
-    public int[] InstructionIndices { get; set; }
+    public ushort[] InstructionIndices { get; set; }
     public GoTo Next { get; set; }
-    public int[] SpeakerIndices { get; set; }
+    public ushort[] SpeakerIndices { get; set; }
     public string Text { get; set; } = string.Empty;
 }
